refactor: move test to inner module
